rename DONATION_PROGRAM --> DONATION_ACCOUNT

+++ src_appaserver/appaserver_purge_temporary_files.sh
@@ -1,29 +1,57 @@
 #!/bin/sh
-# -----------------------------------
-# appaserver_purge_temporary_files.sh
-# -----------------------------------
+# -------------------------------------------------------------------
+# $APPASERVER_HOME/src_appaserver/appaserver_purge_temporary_files.sh
+# -------------------------------------------------------------------
 #
 # Freely available software: see Appaserver.org
-# ---------------------------------------------
+# -------------------------------------------------------------------
 
-delay_minutes=30
-
-if [ -f ./appaserver.config ]
+if [ $# -ne 1 ]
 then
-	appaserver_config="./appaserver.config"
-else
-	appaserver_config="/etc/appaserver.config"
+	echo "Usage: $0 application" 1>&2
+	exit 1
 fi
 
-appaserver_mount_point=`cat $appaserver_config			|
-			grep '^appaserver_mount_point='		|
-			piece.e '=' 1`
+application_name=$1
 
-find	$appaserver_mount_point					\
+delay_minutes=30
+document_root=`get_document_root.e`
+appaserver_data=`get_document_root.e`
+
+directory=${document_root}/${application_name}/appaserver
+
+find	$directory						\
 	-xdev							\
 	-maxdepth 2						\
 	-name '*_[1-9][0-9][0-9][0-9]*'				\
 	-cmin +$delay_minutes					\
 	-exec rm -f {} \; 2>/dev/null
+
+directory=${appaserver_data}
+
+find	$directory						\
+	-xdev							\
+	-maxdepth 2						\
+	-name '*_[1-9][0-9][0-9][0-9]*'				\
+	-cmin +$delay_minutes					\
+	-exec rm -f {} \; 2>/dev/null
+
+#if [ -f ./appaserver.config ]
+#then
+#	appaserver_config="./appaserver.config"
+#else
+#	appaserver_config="/etc/appaserver.config"
+#fi
+#
+#appaserver_mount_point=`cat $appaserver_config			|
+#			grep '^appaserver_mount_point='		|
+#			piece.e '=' 1`
+#
+#find	$appaserver_mount_point					\
+#	-xdev							\
+#	-maxdepth 2						\
+#	-name '*_[1-9][0-9][0-9][0-9]*'				\
+#	-cmin +$delay_minutes					\
+#	-exec rm -f {} \; 2>/dev/null
 
 exit 0
